+++ nrfutil/index.ts
@@ -6,7 +6,7 @@
 
 export { default as prepareSandbox } from './sandbox';
 export { NrfutilSandbox } from './sandbox';
-export type { Progress, SemanticVersion } from './sandboxTypes';
+export type { Progress } from './sandboxTypes';
 export { getNrfutilLogger, setNrfutilLogger } from './nrfutilLogger';
 export {
     getModule,
@@ -14,3 +14,4 @@ export {
     setVerboseLogging,
     getAllModuleVersions,
 } from './modules';
+export { getJlinkCompatibility } from './jlinkVersion';